updated inviewport to use intersection oberserver.
