+++ viewport-example.ts
@@ -5,29 +5,28 @@
  */
 
 /* eslint no-undef: 0 */
-import inViewport from "./inviewport";
+import inViewport from './inviewport';
 
-document.addEventListener("DOMContentLoaded", (): void => {
+document.addEventListener('DOMContentLoaded', (): void => {
   const verticalScrollContainer: HTMLElement | null = document.querySelector(
-    ".vertical-scroll-example"
+    '.vertical-scroll-example'
   );
-
   const horizontalScrollContainer: HTMLElement | null = document.querySelector(
-    ".horizontal-scroll-example"
+    '.horizontal-scroll-example'
   );
-  const shrug: HTMLElement | null = document.querySelector(".shrug");
+  const shrug: HTMLElement | null = document.querySelector('.shrug');
 
   /**
-   * Build Verticle Tiles.
+   * Build Vertical Tiles.
    */
   const setupShrug: Function = (): void => {
     if (shrug) {
-      inViewport(shrug, 0.5, 0.5, [
+      inViewport(shrug, 0.5, [
         () => {
-          shrug.classList.add("visible");
+          shrug.classList.add('visible');
         },
         () => {
-          shrug.classList.remove("visible");
+          shrug.classList.remove('visible');
         },
       ]);
     }
@@ -41,22 +40,22 @@ document.addEventListener("DOMContentLoaded", (): void => {
     container: HTMLElement
   ): void => {
     for (let i = 0; i < 9; i++) {
-      const tile: HTMLElement = document.createElement("div");
+      const tile: HTMLElement = document.createElement('div');
       tile.classList.add(orientation);
       container.appendChild(tile);
 
-      if (orientation === "vertical-tile") {
-        inViewport(tile, "175px", "175px", [
+      if (orientation === 'vertical-tile') {
+        inViewport(tile, 1, [
           () => {
-            tile.classList.add("visible");
+            tile.classList.add('visible');
           },
           () => {
-            tile.classList.remove("visible");
+            tile.classList.remove('visible');
           },
         ]);
       } else {
-        inViewport(tile, 0.5, 0.5, () => {
-          tile.classList.add("visible");
+        inViewport(tile, 0.5, () => {
+          tile.classList.add('visible');
         });
       }
     }
@@ -67,13 +66,13 @@ document.addEventListener("DOMContentLoaded", (): void => {
    */
   const initializeExample: Function = (): void => {
     if (verticalScrollContainer) {
-      buildTiles("vertical-tile", verticalScrollContainer);
+      buildTiles('vertical-tile', verticalScrollContainer);
     }
 
     if (horizontalScrollContainer) {
-      buildTiles("horizontal-tile", horizontalScrollContainer);
-      inViewport(horizontalScrollContainer, 0.01, 0.8, () => {
-        horizontalScrollContainer.classList.add("visible");
+      buildTiles('horizontal-tile', horizontalScrollContainer);
+      inViewport(horizontalScrollContainer, 0.01, () => {
+        horizontalScrollContainer.classList.add('visible');
       });
     }
